2019-04-17 下班提交 shiro整合完毕

+++ butler-common/src/main/java/com/tuxiaoer/shanghai/modules/system/service/RoleService.java
@@ -41,7 +41,7 @@ public interface RoleService {
      * @param userId
      * @return
      */
-    List<Role> getUserAllRole(Integer userId);
+    List<Role> getUserAllRole(Long userId);
 
     /**
      * 保存角色
